+++ apps/recurring_income/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class RecurringIncomeConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "apps.recurring_income"